Fix lexer mode issue when tokenizing dates
Parsing datetime in an array was failing, e.g. `[1993-08-04]`,
`[1993-08-04,1993-08-04]`, etc.

When in lexer mode `DateMode`, there was no rule for matching `]`
resulting in a match to `DateError`.

Additionally, the rule for matching `,` in `DateMode` popped the lexing
mode rather than reverting to `ValueMode`, causing issues when parsing
multiple items in the array due to it eventually popping the mode stack
completely (back to `KeyMode`).

Resolves #4.

Co-authored-by: dauthjan <dipanjan.19932008@gmail.com>