add test case based on real world Cargo.toml
This adds a test input using the collapsed notation that is often
used to specify dependencies in the Rust ecosystem:

    [dependencies]
    derive_more     = { version = "=0.99.17", features = ["deref"] }
    serde_with      = "=2.3.1"
    strum           = { version = "=0.24", features = ["derive"] }

As of 1.1.0 this causes the output to be malformed as the complex
dependencies get interspersed with the simple ones:

    [dependencies]
      [dependencies.derive_more]
        version="=0.99.17"
        features=[
          "deref"
        ]
      serde_with="=2.3.1"
      [dependencies.strum]
        version="=0.24"
        features=[
          "derive"
        ]

Additionally, the indentation of the output is not stable wrt. to
reparsing.