Keep lexer in ValueMode when tokenizing arrays
The lexer uses ValueMode to tokenize a value, and each rule pops the
mode stack after emitting the value token(s). When tokenizing an array,
each value rule now pushes an extra ValueMode to the mode stack, using
`pushValueModeIfInArray()`, ensuring the lexer remains in ValueMode
after that mode stack is popped.

This removes the need for special handling of commas.

Fixes #60.